fix: CodeRabbit レビュー指摘に対応
- 外部URL正規化を src/lib/utils/url.ts に切り出して共通化
- /lives/[id] で href に渡す URL を normalizeExternalUrl で検証
- /lives/[id] の description が空白文字のみの場合は空のカードを描画しないよう trim チェックを追加
